Add controller-level authorization for posts destroying, add more checks to delete specs

+++ app/controllers/posts_controller.rb
@@ -33,6 +33,8 @@ class PostsController < ApplicationController
   end
 
   def destroy
+    authorize(post)
+
     post.destroy
 
     redirect_to root_path
